requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
attrs==23.2.0
av==13.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
deepdanbooru==1.0.0
diffusers==0.30.3
einops==0.8.0
exceptiongroup==1.2.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
google-pasta==0.2.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.62.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.7
imageio==2.34.1
importlib_metadata==8.5.0
importlib_resources==6.4.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.2.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.3
numpy==1.26.4
onnxruntime==1.17.3
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.1
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.2
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qwen-vl-utils==0.0.8
referencing==0.34.0
regex==2024.9.11
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.7
safetensors==0.4.5
scikit-image==0.23.2
scipy==1.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tifffile==2024.4.18
timm==1.0.9
tokenizers==0.20.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.2
transformers==4.45.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.20.2
